Match ComfyUI's arithmetic: bf16 residual stream, fp16 attention, a parity harness
ComfyUI is the reference implementation for these checkpoints, so the two places
the runtime had drifted from it are changed to follow it, and the agreement is
now measured against ComfyUI's own evaluation instead of only our reference.

The residual stream is bf16 and rounds where ComfyUI rounds: the linear output,
the gated product and the sum with the prior state. The GEMM residual epilogues
and the norm prepare kernels read and write bf16, the session copies the caller's
bf16 buffer in and out without casting, and the block ABI goes to 3.

Attention defaults to the fp16 WMMA kernel, which is what ComfyUI's bf16 SDPA
call is closest to on this part; the smoothed int4/int8 QK kernels stay behind
KREA2_ATTN_QK. On ComfyUI's own input to a block that moves the update cosine
from 0.995 to 0.99988, and the 28-block chain from 0.969 to 0.9977, for 5.4 ms
per block.

tools/comfy_step.py dumps a ComfyUI run (every block, the sampler states, the
conditioning, the sigma grid); tools/compare_comfy.py replays the blocks and
tools/compare_comfy_steps.py the whole transformer through the C ABI. Every one
of the 28 blocks agrees to cosine 0.99978 or better on ComfyUI's own input, and
the whole transformer to 0.9981-0.9999 per sampling step.

The hand-written wide down projection goes with its test and benchmark: the
256x128 tile in tools/gen_gemm.py serves every projection now.

+++ reference/loom_ref.py
@@ -69,7 +69,9 @@ class LoomBlocksRef(R.Krea2Ref):
     def block(self, i, x, m, cos, sin):
         p = f"blocks.{i}"
         prescale, preshift, pregate, postscale, postshift, postgate = m.float().unbind(-2)
-        x = x.half()
+        # The residual stream is bf16 with ComfyUI's rounding points: the linear's
+        # output, the gated product, and the sum (the block stack's kernels do the same).
+        x = x.bfloat16()
         normalized = R.rms_norm(x.float(), self.t(f"{p}.prenorm.scale", torch.float32))
         prepared = normalized * (1 + prescale) + preshift
         b, s, _ = x.shape
@@ -81,11 +83,11 @@ class LoomBlocksRef(R.Krea2Ref):
         k = R.apply_rope(R.rms_norm(k.float(), self.t(f"{p}.attn.qknorm.knorm.scale", torch.float32)), cos, sin).half()
         attn = sage_attention(q, k, v).reshape(b, s, -1).half()
         attended = attn.float() * torch.sigmoid(gate.float())
-        x = (x.float() + pregate * self.project(f"{p}.attn.wo.weight", attended)).half()
+        x = (x.float() + (pregate * self.project(f"{p}.attn.wo.weight", attended).bfloat16().float()).bfloat16().float()).bfloat16()
         normalized = R.rms_norm(x.float(), self.t(f"{p}.postnorm.scale", torch.float32))
         prepared = normalized * (1 + postscale) + postshift
         g = self.project(f"{p}.mlp.gate.weight", prepared)
         u = self.project(f"{p}.mlp.up.weight", prepared)
         product = (g * torch.sigmoid(g) * u).half()
         down = self.project(f"{p}.mlp.down.weight", self.rotate_plain(product), rotated=True)
-        return (x.float() + postgate * down).half()
+        return (x.float() + (postgate * down.bfloat16().float()).bfloat16().float()).bfloat16()
